Move the RequestFactory to a separate class and add comments
